Исправления:
- фильтрация городов по наличию ЖД станций
- оставлен только список ЖД станций (для выбора только поездов)
- удалена дата в запросе расписания (для выдачи рейсов на текущую дату и позже)
- добавлен enum DateFormatterHelper для форматирования дат
