+++ TravelSchedule/ViewModel/AgreementViewModel.swift
@@ -8,10 +8,12 @@ import Foundation
 
 @MainActor
 final class AgreementViewModel: ObservableObject, Sendable {
+    // MARK: - Public Properties
     @Published var state: AppState = .content
     @Published var errorType: ErrorType = .serverError
     @Published var usrString: String = ""
     
+    // MARK: - Init
     init() {
         self.usrString = Constants.urlString
     }
