+++ TravelSchedule/ViewModel/StoriesViewModel.swift
@@ -8,11 +8,13 @@ import Foundation
 
 @MainActor
 final class StoriesViewModel: ObservableObject, Sendable {
+    // MARK: - Public Properties
     @Published var state: AppState = .content
     @Published var errorType: ErrorType = .serverError
     @Published var stories: [Stories] = []
     @Published var selectedStory: Int = 0
     
+    // MARK: - Init
     init() {
         getStories()
     }
